Refactorings and Fixings
