+++ collect_app/src/main/java/org/odk/collect/android/database/helpers/XPathDatabaseHelper.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2017 Nafundi
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.odk.collect.android.database.helpers;
 
 import android.database.sqlite.SQLiteDatabase;
@@ -32,9 +16,8 @@ import static android.provider.BaseColumns._ID;
  * This class helps open, create, and upgrade the database file.
  */
 public class XPathDatabaseHelper extends SQLiteOpenHelper {
-    public static final String DATABASE_NAME = "xpath_exprs.db";
-    public static final String XPATH_TABLE_NAME = "xpath_exprs";
-    public static final String INDEXED_EXPRESIONS = "indexed_expressions";
+    public static final String DATABASE_NAME = "xpath_expr_index.db";
+    public static final String XPATH_TABLE_NAME = "xpath_expr_index";
 
     private static final int DATABASE_VERSION = 7;
 
@@ -72,7 +55,8 @@ public class XPathDatabaseHelper extends SQLiteOpenHelper {
         db.execSQL("CREATE TABLE IF NOT EXISTS " + XPATH_TABLE_NAME + " ("
                 + _ID + " integer primary key autoincrement, "
                 + XPathProviderAPI.XPathsColumns.EVAL_EXPR + " text not null, "
-                + XPathProviderAPI.XPathsColumns.TREE_REF + " text );");
+                + XPathProviderAPI.XPathsColumns.GENERIC_TREE_REF + " text, "
+                + XPathProviderAPI.XPathsColumns.SPECIFIC_TREE_REF_ + " text );");
     }
 
 }
