+++ collect_app/src/main/java/org/odk/collect/android/provider/XPathProviderAPI.java
@@ -1,32 +1,16 @@
-/*
- * Copyright (C) 2007 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package org.odk.collect.android.provider;
 
 import android.net.Uri;
 import android.provider.BaseColumns;
 
 public final class XPathProviderAPI {
-    static final String AUTHORITY = "org.odk.collect.android.provider.odk.xpath_exprs";
+    static final String AUTHORITY = "org.odk.collect.android.provider.odk.xpath_expr_index";
 
     private XPathProviderAPI() {
     }
 
     /**
-     * Columns for the XPaths table.
+     * Columns for the XPathExpressionIndex table.
      */
     public static final class XPathsColumns implements BaseColumns {
         private XPathsColumns() {
@@ -35,14 +19,13 @@ public final class XPathProviderAPI {
         /**
          * The content:// style URL for accessing XPaths.
          */
-        public static final Uri CONTENT_URI = Uri.parse("content://" + AUTHORITY + "/xpath_exprs");
+        public static final Uri CONTENT_URI = Uri.parse("content://" + AUTHORITY + "/xpath_expr_index");
 
-        public static final String CONTENT_TYPE = "vnd.android.cursor.dir/vnd.odk.xpath_exprs";
-        public static final String CONTENT_ITEM_TYPE = "vnd.android.cursor.item/vnd.odk.xpath_exprs";
 
-        // These are the only things needed for an insert
+        // These are the only things needed indexing a leaf node which has a value
         public static final String EVAL_EXPR = "evaluationExpression";
-        public static final String TREE_REF = "treeReference";
+        public static final String GENERIC_TREE_REF = "treeReference";
+        public static final String SPECIFIC_TREE_REF_ = "treeReference";
 
     }
 }
